+++ notes/09-controllers/figures/02-use.php
@@ -1,3 +1,3 @@
 <?php
 
-use App\Article;
+use App\Models\Article;