+++ notes/11-one-to-many/figures/08-Articles-use.php
@@ -1,3 +1,3 @@
 <?php
 
-use App\Comment;
+use App\Models\Comment;