+++ notes/17-many-to-many/figures/18-use.php
@@ -1,3 +1,3 @@
 <?php
 
-use App\Tag;
+use App\Models\Tag;